consolidating api and worker dockerfiles
we can use docker multi-stage builds and cut out the need for using
seperate builder dockerfiles and services.  This not only reduces the
amount of configuration, but it also makes the Dockerfiles easier to
read because all the steps are in one file.

Additionally there is no need for production vs development dockerfiles,
since in the docker compose file we can override the entrypoint and
command for local development