need to use || when running optional Docker command
using && means that in some environments the first statement will result
in false and thus the second statement won't be executed.  this also
means that the RUN command as a whole will have a non zero exit code and
cause the Docker build process to fail.  we need to make sure that
regardless of the environment, the RUN command executed has a 0 exit
status