faster loading of synthetic cclf data (#740)
As is, the make load-synthetic-cclf-data target iterates over the ACO size list and issues a docker run command to load the data for each size. By running the docker container once, and looping inside the container execution, we can remove the cost of container setup and teardown with each iteration of the ACO size.