changing error constraint
the ACOS table has 2 constraints; a unique primary key and unique cmsid.
the initial migration created the former constraint before the latter,
thus violations with both will report the former error.  However with
migration 0000015, altering the column deletes, then recreates the
unique cmsid.  thus placing it second, so this test case that violates
both constraints is now going to report an issue with the primary key