BCDA-6858: Refine pull request template (#833)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-6858

## ℹ️ Context for reviewers

I noticed the PR template is very big and daunting. This seems to have
had the effect that folks do not actually fill out most of the template
when opening PRs, even though the notice at the top says not to leave
things blank :)

From speaking with the previous engineers, I noted that several things
were considered optional (e.g. branch naming).

This PR reduces the template size by ~50% and attempts to make it more
friendly and to-the-point. My goal is to reduce the amount of effort
required for creating a pull request, which encourages folks to fill out
the relevant context for reviewers and future team members who may come
across PRs down the line.