BCDA-4456: Remove duplicate calls to SSAS /introspect endpoint (#841)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-4456

## 🛠 Changes

- Rely on ParseToken for token parsing and validation, if it is provided
- Remove validation logic from `RequireTokenAuth` (only check if token
has been pre-validated and stored in the request context)
- Instead of deleting the `AuthorizeAccess` method...
- Merge validation logic from `ParseToken` to `AuthorizeAccess` and
clean up unreachable code

i.e. flow before and after for authorized endpoints:

```
# Before
--> ParseToken (validation logic) 
--> RequireTokenAuth (confirm token exists) 
    --> calls AuthorizeAccess (validation logic)

# After
--> ParseToken
    --> calls AuthorizeAccess (validation logic)
--> RequireTokenAuth (confirm token exists)
```

### New Logic

This PR removes the validation from RequireTokenAuth. If the token is
stored in the request context, _it has already been validated by
ParseToken_ and does not need to be re-validated.

Instead of removing AuthorizeAccess entirely, I have merged the existing
validation logic in `ParseToken` with the logic in `AuthorizeAccess`.
They essentially do the same thing with minor differences.