BCDA-7021: Fix intermittent smoke test failure in patient export E2E test (#844)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7021

## 🛠 Changes

- Update "since" parameter to account for server clock differences

## ℹ️ Context for reviewers

During automated testing, we encounter intermittent 400 request failures
in the "Start Patient export Valid Since" E2E test. This test requests a
patient data export using since=now and confirms that no data is
returned based on the "since" parameter.

Within the application logic, we validate that the "since" parameter is
not in the future. We have identified that this is a potential source of
the intermittent test failure because the "now" time on the Jenkins
server may be slightly ahead of the time on the API server.

To account for this, provide a "since" value with 1 minute of
difference.