BCDA-7081: Increase test timeout for EOB runout export tests (#847)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7081

## 🛠 Changes

- Increase test timeout for EOB runout export jobs from 50 seconds -->
100 seconds

## ℹ️ Context for reviewers

During prod deployment, we ran into two back-to-back failures for the
same job. As part of export job testing, we wait up to 50 seconds for
the job to complete. However, it appears the EOB prod jobs are taking
longer to run than before.

When manually checking the job status afterwards, we see that the job
completed in about 90 seconds. To avoid intermittent smoke test
failures, we should increase the test timeout to allow for this delay in
job processing.

## ✅ Acceptance Validation

I ran the prod smoke tests using this branch, which successfully
completed:
https://management.bcda.cms.gov/job/BCDA%20-%20Test%20-%20Integration%20Smoke/4723/