BCDA-7257: Split opt out utilities into a separate module (#861)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7257

## 🛠 Changes

Create a new Golang module for optout code that can be imported separate
from BCDA code. This includes:
- All related models (SuppressionFile --> OptOutFile, Suppression -->
OptOutRecord, SuppressionFileMetadata --> OptOutFilenameMetadata)
  - Utility function for parsing metadata from filenames
  - Utility function for parsing a single line from an opt out file

All other changes to files are just updating the model names or changing
the model fields from private to public (e.g. `suppression.fileID` -->
`suppression.FileID`) since the models are now in a separate module.

For reviewers, I would recommend skimming the existing files and
focusing mainly on the new `optout` folder 🙂

## ℹ️ Context for reviewers

AB2D and DPC are currently in the process of automating data-sharing
opt-out requests. It is expected that the file formats will be
consistent with the existing files that BCDA receives, so a lot of the
existing Go code is reusable.

DPC is likely going to follow the pre-established pattern of building
lambdas in Go, so we expect that BCDA's code could be directly used as
part of the lambda. This PR pulls out generic utilities that are not
related to BCDA's implementation in any way (e.g. reading from the local
filesystem or saving to the BCDA database.)

## ✅ Acceptance Validation

- Unit tests continue to pass
- Was able to create a new local go program and import these utilities
using `go get
github.com/CMSgov/bcda-app/output@777710bbe26fba867d10a61a84f68a1b1fdb1f76`.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.