BCDA-7135: bcda api logging improvements (#865)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7135

## 🛠 Changes

- added new middleware in logging.go that adds request ID and the cms ID
to the context for next handler
- updated requests.go logging statements to utilize logger that is
passed via context

## ℹ️ Context for reviewers

Updating the request processing to add a logger to the context via
middleware. Request ID and CMS ID can be used in future logging calls
and new fields can be added to the requests context logger as needed.

Looked into database logging to possibility be included in this PR, but
database is using the default `logrus.logger`. Creating a separate
ticket to address that.

## ✅ Acceptance Validation

Current tests pass, new test added, and ran locally to check logs &
context values.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.