BCDA-7135: bcdaworker logging improvements (#869)
…worker uses ctx to pass logger

## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7135

## 🛠 Changes

- The bcda worker.go and que.go will utilize context to create and pass
around log fields.
- ields cmsID and jobID have been renamed to cms_id and job_id to for
consistency and querying
- log strings have been modified so that field values are in their own
individual fields instead of the free text msg field
- common logging components have been moved to log/logger.go and import
names updated
- add a single `logger.info()` in `requests.go/bulkrequest()` so that
job ID and request ID can be correlated

## ℹ️ Context for reviewers

The BCDA worker does not use the initial request context when processing
jobs. It does use the Job ID, which is a common identifier that we can
use to query events on. When we use the request_id in splunk we can view
related events for the job and use the job ID to query events further.
This requires the addition of one informational log event when the job
ID is created.

The que.go file has been modified so that when we create our context for
the job processing, it uses it to pass around log fields to other
function calls. This has eliminated the need to declare the fields
repeatedly when we use WithFields({...}) and removed redundant or
unnecessary log calls.

Below is a new log entry (local example) that will be generated with
each request that will contain both the request ID and the job ID, so
that events can be correlated.

<img width="1248" alt="image"
src="https://github.com/CMSgov/bcda-app/assets/126501259/1dad818f-3d79-4b21-b2e1-29e638eb2abd">



## ✅ Acceptance Validation

Tests updated and passing, tested app running locally.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.