BCDA-7126 BCDA: Improve coverage for "bcda/suppression/suppression.go" (#871)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7126

## 🛠 Changes

Removed outdated package in suppression.go (ioutil)
Updated unit tests for suppression.go to get to 95.2% statement
coverage.

## ℹ️ Context for reviewers

In addition to the new tests, of which coverage is primarily added by
implementing tests for the ImportSuppressionDirectory function, existing
tests are modified to utilize UTC as the standard date format in case
someone runs the tests locally (instead of on docker, where the default
time is UTC). Future improvements to coverage on this file can be made
by implementing a means to mock the repository. This would enable
leaving only ~2 statements uncovered.

## ✅ Acceptance Validation


![image](https://github.com/CMSgov/bcda-app/assets/120701369/b75050f6-4a0d-441a-ac2b-3722868f5744)

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.

---------

Co-authored-by: Alex Dzeda <alex@e14s.com>