BCDA-6594: Improve logging when receiving BFD 400/500 status codes (#882)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-6594

## 🛠 Changes

Using struct `JobEnqueueArgs` in functions to replace individual
arguments of `cmsID` and `jobID`, so that we can include those data
points in our requests to BFD. The field of `cmsID` has been added to
the model `JobEnqueueArgs`. I ended up the available struct with the
additional field because many of the functions that required a changed
needed some field data from JobEnqueueArgs. This reducing the number of
parameters that functions expect and (hopefully) improves readability.

## ℹ️ Context for reviewers

Our BB client has two methods `GetMetaData` and
`GetPatientHashIdentifier` that do not include the job ID or CMS ID in
the request headers, which results in those values being empty when the
request is logged.

## ✅ Acceptance Validation

Wrote new test and modified existing; all are passing.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.