BCDA-7480: Add alerting when CCLF-8 import job fails to open a file (#883)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7480

## 🛠 Changes

Updated the cli command `import-cclf-directory` to return the error when
one occurs. If no errors come from the `importcclfdirectory` function in
cli.go but a file is skipped or fails import, then a new error is
created and returned to be handled by the jenkins script.

## ℹ️ Context for reviewers

Jenkins pipeline currently always runs successfully even if cclf files
are skipped or fail to import.

## ✅ Acceptance Validation

Tests still passing. This jenkins pipeline does not have the option to
run from a development branch, for both the pipeline and the app. This
will need to be deployed and a mock test file will need to be created in
order to test.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.