BCDA-7550: Job ID and CMS ID missing from BFD request/response (#887)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7550

## 🛠 Changes

- Modified BBRequestSuite to configure a logger for the client so tests
can be run individually
- Modified createQueueJobs to set the CMS ID in the job arguments
- Added a test in bluebutton_test.go to make sure we are getting values
set when making a request to BFD

## ℹ️ Context for reviewers

Requests outgoing to BFD are missing values for the CMS_ID and the Job
ID when attempting to get the Patient Identifier.

## ✅ Acceptance Validation

Tested changes locally and with unit tests. Confirmed that 

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.