BCDA-7480: Update errors returned from import-cclf (#885)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7480

## 🛠 Changes

Updating `import-cclf` command to return an error when at least one file
has failed to be imported or a file has been skipped, or if the
successfully imported file count is 0.

Smoke tests are failing because the suppression files get added to the
skipped count, if they exist in the directory. Since we don't want false
positives, a check has been added to see if the file we are currently
importing is an opt out file. If it is, we ignore it altogether and
don't increment the count of the files skipped.

## ℹ️ Context for reviewers

Smoke tests in dev are failing because opt-out files are being added to
the skipped count, which causes the synthetic EFT import to fail.

## ✅ Acceptance Validation

modified tests that import cclf files to also check for opt out files
and not add to skipped count.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.