BCDA-7134: Update error logging (#888)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7134

## 🛠 Changes

- Add new error logs when returning errors to the API caller
- Add new error/warning logs when returning from a function in response
to an error or other conditional scenarios
- Ensure errors are wrapped (`err = errors.Wrap(err, "my message"))`
instead of ignored (`err = errors.New("my message")`)
- Surface error from `HashIdentifier`
- Use ctx logger when plausible
- Use `error` level when logging 500 errors

## ℹ️ Context for reviewers

In several areas of the codebase, we are running into errors and taking
action on them without logging the original error. We should ensure that
errors are logged when possible, and that we are otherwise logging
contextual information when making conditional decisions (e.g. in the
`hasDuplicates` function, we should be logging each conditional check.)

I basically looked at every instance of `err != nil` and verified
whether it was being logged or returned to a downstream caller (who
should also be checking for err != nil, and then logging it or returning
it farther downstream.)

Note that I did not make any changes to the writer files to utilize the
ctx logger since they were heavily used and would likely require more
involved changes there.

## ✅ Acceptance Validation

Unit Testing.