BCDA-7655: Update BCDA database settings (#903)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7655

## 🛠 Changes

Updated maximum number of DB connections from 40 -> 60
Added idle connection timeout

## ℹ️ Context for reviewers

There are inefficiencies in connection handling for the go postgres /
sql libraries that we use, and these settings should help mitigate those
inefficiencies by pruning idle connections. Specifically, the pgx
library doesn't check if a connection is active before trying to use it,
and an aggressive idle timeout should reduce the number of instances
where a stale connection is thought to be available.
I also updated some default configs for timeouts for worker-> BB + app->
SSAS to reflect the configs that are actually deployed. In the case of
app -> SSAS, the timeout was 1/10 what we actually deploy as an env var,
and BB was 1/20.

## ✅ Acceptance Validation

Manual verification of information in the commit

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.

---------

Co-authored-by: Alex Dzeda <alex@e14s.com>