BCDA-7864: Add Sonarqube quality gate check to PRs (#929)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7864

## 🛠 Changes

- Add Sonarqube quality gate to pull request check

## ℹ️ Context for reviewers

We run Sonarqube quality gates as part of builds and deployments, but
not as part of PRs, so sometimes we'll merge a change and break the
build/deploy.

We also have some security checks that we manage manually at the moment
using a scheduled task. We can automate this within the pipeline.

## ✅ Acceptance Validation

See checks.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.