BCDA-7944: Requests to BFD are external in New Relic (#931)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7944

## 🛠 Changes

- HTTP requests to BFD are instrumented using `StartExternalSegment`
- Transactions started for HTTP requests to BFD are no longer marked as
a web transaction


## ℹ️ Context for reviewers

We have been marking calls to BFD as internal, when they are an external
entity. We now have an external segment started when the request to BFD
starts. In the future, we should start the transaction right as the job
starts processing in que.go and segment the transaction as appropriate.
When the job has finished processing, we can stop the transaction. This
will give more accurate traces in new relic and we can get a better idea
of how our job processing is performing.

## ✅ Acceptance Validation

Deployed to TEST env, will post screenshots and details internally.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.