BCDA-8120: Parse Accept-Encoding to search for 'gzip' (#940)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8120

## 🛠 Changes

Update gzip logic to properly identify a gzip header

## ℹ️ Context for reviewers

>90% of requests to BCDA pass along a valid Accept-Encoding including
gzip, but the current method of parsing is only recognizing "gzip" if
that header exactly equals "gzip". This PR fixes that and searches for
"gzip" and becomes agnostic to whatever else accompanies in that header.

## ✅ Acceptance Validation

Added unit tests showing examples of headers we see in production

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.