BCDA-8062: OpenAPI Specification Review (#942)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8062

## 🛠 Changes

**Updates**:

- organized endpoints by noun
- removed alr endpoints, since they are not in use
- moved _auth under auth tag instead of metadata
- removed http 500 responses, as those are generally unexpected and
there's no action items for customers to take
- updated endpoint descriptions and summaries to have less ambiguous
language

**Findings:**

- all v1 and v2 endpoints except Patient and Group have no difference in
functionality. ie, attribution, jobs.
- we use a mix of `snake_case` and `camelCase` for our path and query
params, and some of our endpoints nouns use a mix of upper and lower
case.
- I found a tool called schemathesis that does OpenAPI contract
validation for APIs. The tool is very useful; you can configure lots of
different options for testing and it's language agnostic. Extending the
functionality would have to be written in python, though. I used this
for some basic validation of the spec.
- we can parameterize the swagger docs to use a different base path,
depending on our env that we are deploying to. Then each endpoint won't
need the api/v#/... prefix

## ℹ️ Context for reviewers

Our OpenAPI docs are out of date and may not match the actual behavior
of the API. This review is meant to reasonably update the endpoints,
descriptions, responses, etc. Validation and contract testing will come
in a later ticket.

## ✅ Acceptance Validation

TBD, needs team review and approval. 

For a comparison of the current spec:
https://sandbox.bcda.cms.gov/api/v1/swagger/

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.