BCDA-8154: Update capabilitystatement operation names (#945)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8154

## 🛠 Changes

Updated Operation names to meet FHIR spec

## ℹ️ Context for reviewers

Names of operations in a CapabilityStatement should refer to,
conceptually, an OperationDefinition code. In the case of an operation
such as Patient/$export for the bulk FHIR IG, the parameter used to
instantiate the operation is export, so patient-export should be changed
to export.
There is also some odd auto-linting with this, which appears to go away
after this is merged?

## ✅ Acceptance Validation

Updated + passed unit tests. 

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.