BCDA-8146: Make file server content-encoding aware (#948)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8146

## 🛠 Changes

Added magic number check for files, to determine if it is gzip encoded
(or not).

## ℹ️ Context for reviewers

As part of BCDA's gzip epic, we are aiming to simultaneously decrease
costs on EFS, while also increasing the overall throughput of files.
When we switch over from primarily storing content uncompressed to
primarily storing data gzip-encoded, there will be at least some period
of time where both encodings are present. The BCDA application will now
check the magic number
(https://en.wikipedia.org/wiki/Magic_number_(programming)) to determine
if content is encoded or not, and will handle the data appropriately
(decompressing the file if an Accept-Encoding header of gzip is not
present).

There is a binary file,
shared_files/gzip_feature_test/corrupt_gz_file.ndjson, that is not able
to be opened with a normal text editor. It's advised to open files of
this type in a hex editor for manual inspection. In this case, the file
has the magic number of a gzip file, but the subsequent data is not
compliant.

## ✅ Acceptance Validation

Added unit tests to handle the various cases. 

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.