BCDA-8104 (1/2): Add lambda entrypoint for running opt out import (#947)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8104

## 🛠 Changes

- Translates some parameter store utilities from DPC into a `bcdaaws`
package
- Adds a new `bcda/lambda` directory for housing lambda entrypoints
- Added lambda entrypoint for opt out import
- Updated database connection initializer to support loading values from
Parameter Store

## ℹ️ Context for reviewers

This PR allows us to run the opt out import functionality as a lambda. A
few points to note related to this:

### 1. Loading database configuration

The database connection usually requires DATABASE_URL and
QUEUE_DATABASE_URL to be set as environment variables before the
application is started. For lambda applications, secret values need to
be pulled at runtime within the application.

Originally, I wanted to avoid using the `connection.go` functionality at
all. However, it appears that the init function always runs if the
`database` package is imported at any point in the import chain. Because
of this, the lambda entrypoint was always looking for those env vars and
then crashing.

Instead, I've added a fallback to the initializer to check parameter
store if the ENV var is set (which it will be, for lambda.)

### 2. Lambda environment variables

Lambda is configured differently from our EC2 applications. In
particular, we are using shared infrastructure modules and only have
these environment variables available to us:


https://github.com/CMSgov/ab2d-bcda-dpc-platform/blob/f9fa64bde4bf3399d8cbfdc75e0eff0f23697b5a/terraform/services/opt-out-import/main.tf#L53-L57

When retrieving the environment, we need to use ENV instead of
DEPLOYMENT_TARGET.

### 3. Command line arguments

Lambdas receive string-y events as input. With the shared opt-out import
infrastructure, they will be receiving S3 event notifications through
SNS for any new files dropped into the bucket directory.

Since we do not have the usual CLI arguments for the BFD assume role arn
and BFD import bucket, those values are pulled from new places:

- BFD assume role ARN: Already available in parameter store, so pull it
from there.
- BFD import bucket: Can be parsed from the S3 event notification when
it comes in.

## ✅ Acceptance Validation

- Unit Testing

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.