BCDA-8104 (2/2): Add lambda entrypoint for CCLF import (#949)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8104

## 🛠 Changes

- Add lambda entrypoint for CCLF import
- Fix bug in setting filepath

## ℹ️ Context for reviewers

This PR allows us to run the CCLF import functionality as a lambda. This
should look _very_ similar to the Opt Out Import entrypoint, with some
minor changes to run the CCLF import instead of Opt Out import.

In particular...

1. We expect the infrastructure to be set up similarly (Lambda is
triggered by new files arriving in S3)
2. Lambda code should iterate over all files in the S3 destination and
look for CCLF files to import

Note that this PR builds on top of #947. I will wait to merge this until
947 is merged.

## ✅ Acceptance Validation

- Unit testing