Ensure temp directory is cleared on process start (#958)
## 🎫 Ticket

N/A

## 🛠 Changes

Ensure that the temporary directory is emptied upon process start.

## ℹ️ Context for reviewers

After changing the worker's logic to write to EBS volumes prior to
writing compressed files to EFS, some monitoring showed that if the bcda
worker process exits in a non-graceful manner, data could accumulate on
the EBS volume and cause out of space errors. This PR ensures that when
the process starts, the temporary directory is cleared.

## ✅ Acceptance Validation

Added unit tests that pass. 

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.