Increase export job waiting period for smoke tests (#967)
## 🎫 Ticket

N/A

## 🛠 Changes

- Update maxRetries from 10 --> 20 everywhere
- Autoformat postman collection JSON files (oops sorry..)

## ℹ️ Context

After https://github.com/CMSgov/bcda-app/releases/tag/r226 was released,
it looks like jobs are taking a _little_ bit longer to run in
production. It's not clear if it's because of something in the release
or tangential.

This updates maxRetries everywhere (previously, a handful of maxRetries
were already updated to 20 due to intermittent issues.)

## 🧪 Validation

Successful smoke test on production with r226 after a series of
intermittent failures.