Remove stdout print statements (#970)
## 🎫 Ticket

N/A

## 🛠 Changes

Remove print statements

## ℹ️ Context

These were causing the error alarm to go off. We used to duplicate
messages to stdout and to the formatted JSON Logger since only stdout
would be printed in Jenkins logs. Since we are running in lambda now, we
can remove that.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Unit Testing