BCDA-8215: Remove completed job count application logic (#964)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8215

## 🛠 Changes

- Remove all application logic referencing completed job count

## ℹ️ Context

This column, by it's own commented definition, is not a reliable
indicator of the number of actually completed sub-jobs for an export
job. It appears that its only use is in the in-progress job status
header to indicate roughly how far along the job processing is.

Instead of attempting to maintain this field, we should be relying on
the source of truth (the number of completed job keys). To avoid
confusion and to reduce code complexity, we are removing this column.

Before removing the column from the database, we must stop reading it
from the application.

## 🧪 Validation

- Unit testing