BCDA-7900: Add integration test validation for opt-out import (#963)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7900

## 🛠 Changes

Integration testing for opt-out lambda has had additional steps added:
- queries the database to see if the opt out file was ingested
- queries the database to see if suppression entries were created from
the ingested file

## ℹ️ Context

Integration testing is required to ensure that future changes will not
break existing functionality and that the opt-out lambda is working as
expected.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Ran using happy and unhappy paths. Validated that sensitive data is not
printed to stderr or stdout in the event of an issue, and that the file is actually ingested into application database.