BCDA-8162: Update cli commands to be more inclusive. (#973)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8162

## 🛠 Changes

Updated cli references + models to blacklist, updated to denylist.

## ℹ️ Context

As part of making BCDA's code more inclusive, we are removing the terms
"blacklist" and "whitelist."

## 🧪 Validation

Smoke tests + unit tests passing. Note, there is going to be an
associated PR to the jenkins job for the denylist that references the
new cli command name.